working through hacks
