+++ src/pandas_profiling/model/correlations.py
@@ -13,8 +13,6 @@ from scipy import stats
 
 from pandas_profiling.config import config
 
-
-
 def cramers_corrected_stat(confusion_matrix, correction: bool) -> float:
     """Calculate the Cramer's V corrected stat for two variables.
 
@@ -67,12 +65,10 @@ def categorical_matrix(
     Returns:
         A correlation matrix for categorical variables.
     """
-    # TODO: hack
-    from pandas_profiling.model.typeset import Category
     categoricals = {
         column_name: df[column_name]
         for column_name, variable_type in variables.items()
-        if variable_type == Category
+        if variable_type.categorical
         # TODO: solve in type system
         and config["categorical_maximum_correlation_distinct"].get(int)
         >= df[column_name].nunique()
@@ -213,25 +209,8 @@ def perform_check_correlation(
     Returns:
         The variables that are highly correlated.
     """
+    cols = correlation_matrix.columns
+    bool_index = abs(correlation_matrix.values) >= threshold
+    np.fill_diagonal(bool_index, False)
+    return {col: cols[bool_index[i]].values.tolist() for i, col in enumerate(cols) if any(bool_index[i])}
 
-    corr = correlation_matrix.copy()
-
-    # TODO: use matrix logic
-    # correlation_tri = correlation.where(np.triu(np.ones(correlation.shape),k=1).astype(np.bool))
-    # drop_cols = [i for i in correlation_tri if any(correlation_tri[i]>threshold)]
-
-    mapping: Dict[str, List[str]] = {}
-    for x, corr_x in corr.iterrows():
-        for y, corr in corr_x.iteritems():
-            if x == y:
-                break
-
-            if corr >= threshold or corr <= -1 * threshold:
-                if x not in mapping:
-                    mapping[x] = []
-                if y not in mapping:
-                    mapping[y] = []
-
-                mapping[x].append(y)
-                mapping[y].append(x)
-    return mapping
